[r] Refractor CM integration test
